+++ util/helper/camera.go
@@ -1,0 +1,7 @@
+// Copyright 2016 The G3N Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+package helper
+
+// TODO: draw the camera frustum
